requirements.txt
django==1.10.6
djangorestframework==3.6.0
psycopg2==2.7.1
markdown==2.6.8
django-simple-history==1.8.2